requirements.txt
transformers==4.28.1
sentencepiece==0.1.98
underthesea==6.2.0
gensim==4.3.1
numpy==1.24.2
pandas==2.0.0
matplotlib==3.7.1
torch==2.0.0
ipykernel==6.22.0
openpyxl==3.1.2
python-dotenv==1.0.0
pydantic==4.2.0
google-cloud-vision==3.4.1
unidecode==1.3.6
uvicorn==0.22.0
fastapi==0.95.1